requirements.txt
pillow
keyboard
pure-python-adb
uiautomator2
google-generativeai
sentence-transformers
